requirements.txt
flask==2.0.2
transformers==4.12.0
faiss-cpu==1.7.1
sentence-transformers==2.1.0
